Ruby 1.8.7/Mocha issue: stub out :freeze method to prevent teardown bugs.
If, in response to a delete/destroy call, a CassandraMapper::Base instance
is marked as :frozen? during a test, Mocha will choke and break the test run
when trying to remove stubbed methods from that object.  This commit tweaks
things to prevent actual freezing (by stubbing out the :freeze method) or by
using expectations rather than assertions when verifying that an object is
frozen by the deletion logic.